68 implement saved cities feature

+++ backend/src/main/java/com/nemetabe/solarwatch/model/dto/member/MemberDto.java
@@ -1,8 +1,8 @@
 package com.nemetabe.solarwatch.model.dto.member;
 
 import com.nemetabe.solarwatch.model.entity.Member;
-
-public record MemberDto(Integer id, String name, String email) {
+//TODO
+public record MemberDto(Long id, String name, String email) {
     public MemberDto(Member member) {
         this(member.getId(), member.getName(), member.getEmail());
     }
